pyproject.toml
[project]
name = "import-cam"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datetime",
    "Image",
    "Pillow",
]
